Add files via upload
Arduino code to use an Arduino Leonardo or equivalent with an Adafruit Trellis I2C keypad as an HID (appears as USB keyboard) for logging astronomy observations. Keybindings and pins used are documented in the code comments. Keybindings are mostly issuing one or more characters for a Trellis keypad press but a few are custom key sequences for a Linux ( <alt><tab>), Emacs, and Kstars environment. Led brightness is hard coded. A later version that reads a potentiometer to set led brightness seems to have been misplaced.